Rather, with its future -- see also Skyyrunner#1.

+++ veh_interact.cpp
@@ -1021,7 +1021,7 @@ void veh_interact::countDurability()
 
         if(part.hp < part_dur)
         {
-            double damageRatio = part.hp / part_dur;
+            double damageRatio = (double) part.hp / part_dur;
             if (!isnan(damageRatio) && (damageRatio < mostDamaged))
             {
                 mostDamaged = damageRatio;
